requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.2
cryptography==45.0.5
dnspython==2.7.0
dotenv==0.9.9
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.110.1
h11==0.16.0
idna==3.10
motor==3.3.1
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pymongo==4.5.0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.9
razorpay==1.4.2
requests==2.32.4
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.29.0
